feat: automated proxy contract deployment, updated tests
